Move Regex classes into PHPStan\Type\Regex namespace
