+++ src/Type/Php/RegexArrayShapeMatcher.php
@@ -14,6 +14,10 @@ use PHPStan\Type\Constant\ConstantIntegerType;
 use PHPStan\Type\Constant\ConstantStringType;
 use PHPStan\Type\IntegerRangeType;
 use PHPStan\Type\IntegerType;
+use PHPStan\Type\Regex\RegexAlternation;
+use PHPStan\Type\Regex\RegexCapturingGroup;
+use PHPStan\Type\Regex\RegexExpressionHelper;
+use PHPStan\Type\Regex\RegexGroupParser;
 use PHPStan\Type\StringType;
 use PHPStan\Type\Type;
 use PHPStan\Type\TypeCombinator;
