+++ tests/PHPStan/Rules/Regexp/RegularExpressionPatternRuleTest.php
@@ -4,7 +4,7 @@ namespace PHPStan\Rules\Regexp;
 
 use PHPStan\Rules\Rule;
 use PHPStan\Testing\RuleTestCase;
-use PHPStan\Type\Php\RegexExpressionHelper;
+use PHPStan\Type\Regex\RegexExpressionHelper;
 use function sprintf;
 use const PHP_VERSION_ID;
